+++ CHANGELOG.md
@@ -2,6 +2,17 @@
 
 ## [Unreleased]
 
+## [0.0.3]
+### Added
+- Added Ride file creation from lookup
+- Added reference contributor for functions in one file
+- Added find usages for functions in one file
+- Added completion for user-defined functions
+- Added completion for user-defined variables
+
+### Fixed
+- Language grammar errors
+
 ## [0.0.2]
 ### Added
 - Built-in structures, types and functions completion